[GH-314] Persist user settings in native Linux app
This commit adds persistence for the user settings in the native Linux
app. The settings are written to $XDG_CONFIG_HOME/focalboard/settings.
If XDG_CONFIG_HOME is unset, the app falls back to $HOME/.config.

The change hooks into the existing settings export already used in the
nativve macOS app which means the settings are persisted immediately on
change.

Unfortunately, the golang webview package uses a custom native binding
technology and doesn't allow to define WebKit message handlers. As a
result, a dedicated message handler function was added to the existing
NativeApp object. In the native macOS app, said handler is
short-circuited to window.webkit.messagehandlers.[NAME].postMessage.
This has the benefit that the web app remains agnostic of the particular
native binding mechanism.

Relates to: #314