Create a Github Action based Docker Publisher
This PR creates a simple Github Action Workflow that run Docker and uploads to Dockerhub. `DOCKERHUB_USERNAME` and `DOCKERHUB_TOKEN` are required for the account name and API token. This will also generate a cross platform build for arm64/v8 and armv7 so that we can run it on Raspberry Pi/Oracle Cloud/Graviton2 instances easily.