Close migrations
This was the original reason behind focalboard
not working with the DB RPC API.

Because we were not closing the migrations instance,
the original connection was never being
released and therefore we ran out of connections.

The fix we did was kind of a workaround
because we would open a new DB instance and then close
it explicitly which would close the connection also,
but that meant we could not reuse the connection.

Closing the driver also closes the source driver
which was also something that wasn't being done.

https://community-daily.mattermost.com/boards/workspace/zyoahc9uapdn3xdptac6jb69ic/285b80a3-257d-41f6-8cf4-ed80ca9d92e5/495cdb4d-c13a-4992-8eb9-80cfee2819a4/cxg16f8nro7bhxjdwpnqdnumj5o