Update .gitignore (#3563)
When switching back and forth between main and release-7.2 it is easy to accidentally check in the go.work files, which are not included in the 7.2 branch .gitignore.  This PR is a safe change that simple adds the go.work files to the .gitignore file to prevent those files getting checked into this branch.