Do not shortcircuit the schema table populate process
The order in which the migration slice is sorted is dependant on the
`ReadDir` implementation of the filesystem. It should come back sorted
by name, but I'm adding this patch just to be sure we register all the
required migrations instead of shortcircuiting early if the order is
not the one we're expecting.