📝 docs(CONTRIBUTING.md): Update contribution guidelines due to change in project maintenance
The contribution guidelines have been updated to reflect the recent changes in the project's maintenance. As of September 15th, 2023, Mattermost, Inc. staff are no longer reviewing or merging pull requests for either Focalboard or the Mattermost Boards plugin in this repository. The community is encouraged to fork this repository for continued development and contributions. This change is due to Mattermost's decision to focus developer resources on improving the platform’s performance and core features. The list of contributors has been updated to 'Past maintainers'.